html/dom/render-blocking/parser-inserted-modulepreload-link.tentative.html

<!DOCTYPE html>
<title>Parser-inserted modulepreload links with "blocking=render" are render-blocking</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/test-render-blocking.js"></script>

<link rel="modulepreload" blocking="render" href="support/dummy-1.mjs?pipe=trickle(d1)">
<div id="dummy">Lorem ipsum</div>

<script>
test_render_blocking(
  document.querySelector('link'),
  async () => {
    // Attach the module script into document. It should load instantly due to preloading.
    const script = document.createElement('script');
    const scriptObserver = new LoadObserver(script);
    script.type = 'module';
    script.src = 'support/dummy-1.mjs?pipe=trickle(d1)';
    const beforeLoad = performance.now();
    document.body.appendChild(script);
    await scriptObserver.load;
    const afterLoad = performance.now();
    assert_equals(document.getElementById('dummy').textContent, '1',
                  'preloadedmodule script should be executed');
    assert_approx_equals(beforeLoad, afterLoad, epsilon,
                         'preloaded module script should load instantly');
  },
  'Parser-inserted render-blocking module script is preloaded and evaluated');
</script>
